requirements.txt
torch==2.1.0
transformers==4.32.0
datasets==3.2.0
graph-walker==1.0.6
sentencepiece==0.1.99
tqdm==4.67.1
pybind11==2.11.1
networkx==3.1
peft==0.5.0
python-dotenv==1.0.1
tiktoken==0.8.0
numpy==1.26.4
protobuf==4.25.5
uvicorn==0.34.0
pydantic==2.10.6
fastapi==0.115.11
sse_starlette==2.2.1
trl==0.9.4